[google-ti-feeds] refacto for better usecases/mvp separation
