+++ external-import/google-ti-feeds/connector/src/custom/configs/malware/fetcher_config_malware.py
@@ -1,0 +1,44 @@
+"""GTI fetcher configurations for malware families.
+
+This module defines configurations for fetching malware family entities
+from the Google Threat Intelligence API using the generic fetcher system.
+"""
+
+from connector.src.custom.exceptions import (
+    GTIMalwareFetchError,
+)
+from connector.src.custom.models.gti.gti_malware_family_model import (
+    GTIMalwareFamilyData,
+    GTIMalwareFamilyResponse,
+)
+from connector.src.utils.fetchers.generic_fetcher_config import GenericFetcherConfig
+
+GTI_MALWARE_FETCHER_CONFIG = GenericFetcherConfig(
+    entity_type="malware_families",
+    endpoint="/collections/{entity_id}",
+    display_name="malware families",
+    display_name_singular="malware family",
+    exception_class=GTIMalwareFetchError,
+    response_model=GTIMalwareFamilyData,
+    method="GET",
+    headers={"accept": "application/json"},
+    timeout=60.0,
+    response_key="data",
+)
+
+GTI_MAIN_MALWARE_FAMILY_FETCHER_CONFIG = GenericFetcherConfig(
+    entity_type="malware_families",
+    endpoint="/collections",
+    display_name="Malware Families",
+    exception_class=GTIMalwareFetchError,
+    response_model=GTIMalwareFamilyResponse,
+    method="GET",
+    headers={"accept": "application/json"},
+    timeout=60.0,
+    response_key=None,
+)
+
+MALWARE_FETCHER_CONFIGS = {
+    "malware_families": GTI_MALWARE_FETCHER_CONFIG,
+    "main_malware_families": GTI_MAIN_MALWARE_FAMILY_FETCHER_CONFIG,
+}
