+++ external-import/google-ti-feeds/connector/src/custom/configs/malware/gti_config_malware.py
@@ -1,0 +1,27 @@
+"""GTI configuration for malware families.
+
+This module defines configuration settings specific to GTI malware family imports.
+"""
+
+from typing import List
+
+from connector.src.custom.configs.gti_config_common import (
+    ALLOWED_ORIGINS,
+    GTIBaseConfig,
+    validate_origins_list,
+)
+from pydantic import field_validator
+
+
+class GTIMalwareConfig(GTIBaseConfig):
+    """Configuration for GTI malware family imports."""
+
+    malware_family_import_start_date: str = "P1D"
+    import_malware_families: bool = False
+    malware_family_origins: List[str] | str = "All"
+
+    @field_validator("malware_family_origins", mode="before")
+    @classmethod
+    def split_and_validate_malware_family_origins(cls, v: str) -> List[str]:
+        """Split and validate a comma-separated string into a list and validate its contents."""
+        return validate_origins_list(v, "malware family origin", ALLOWED_ORIGINS)
