+++ external-import/google-ti-feeds/connector/src/custom/orchestrators/__init__.py
@@ -4,8 +4,23 @@ This package contains orchestrators that coordinate the fetching, conversion,
 and batch processing of data from the Google Threat Intelligence API.
 """
 
+from connector.src.custom.orchestrators.malware.orchestrator_malware import (
+    OrchestratorMalware,
+)
+from connector.src.custom.orchestrators.report.orchestrator_report import (
+    OrchestratorReport,
+)
+from connector.src.custom.orchestrators.threat_actor.orchestrator_threat_actor import (
+    OrchestratorThreatActor,
+)
+
+from .base_orchestrator import BaseOrchestrator
 from .orchestrator import Orchestrator
 
 __all__ = [
+    "BaseOrchestrator",
     "Orchestrator",
+    "OrchestratorMalware",
+    "OrchestratorReport",
+    "OrchestratorThreatActor",
 ]
