+++ external-import/google-ti-feeds/connector/src/custom/orchestrators/orchestrator.py
@@ -5,25 +5,25 @@ using the proper fetchers/converters/batch processor pattern.
 """
 
 import logging
-import re
 from typing import Any, Dict, Optional
 
-from connector.src.custom.client_api import ClientAPI
-from connector.src.custom.configs import (
-    MALWARE_FAMILY_BATCH_PROCESSOR_CONFIG,
-    REPORT_BATCH_PROCESSOR_CONFIG,
-    THREAT_ACTOR_BATCH_PROCESSOR_CONFIG,
-    GTIConfig,
+from connector.src.custom.configs import GTIConfig
+from connector.src.custom.orchestrators.malware.orchestrator_malware import (
+    OrchestratorMalware,
 )
-from connector.src.custom.convert_to_stix import ConvertToSTIX
+from connector.src.custom.orchestrators.report.orchestrator_report import (
+    OrchestratorReport,
+)
+from connector.src.custom.orchestrators.threat_actor.orchestrator_threat_actor import (
+    OrchestratorThreatActor,
+)
 from connector.src.octi.work_manager import WorkManager
-from connector.src.utils.batch_processors import GenericBatchProcessor
 
 LOG_PREFIX = "[Orchestrator]"
 
 
 class Orchestrator:
-    """Orchestrator for fetching and processing data."""
+    """Main orchestrator that delegates to specialized orchestrators."""
 
     def __init__(
         self,
@@ -47,136 +47,40 @@ class Orchestrator:
         self.tlp_level = tlp_level.lower()
 
         self.logger.info(f"{LOG_PREFIX} API URL: {self.config.api_url}")
-        self.logger.info(
-            f"{LOG_PREFIX} Report import start date: {self.config.report_import_start_date}"
-        )
-        self.logger.info(
-            f"{LOG_PREFIX} Threat actor import start date: {self.config.threat_actor_import_start_date}"
-        )
+        self.logger.info(f"{LOG_PREFIX} Initializing orchestrator")
 
-        self.client_api = ClientAPI(config, logger)
-        self.converter = ConvertToSTIX(config, logger, tlp_level)
-        self.report_batch_processor = self._create_report_batch_processor()
-        self.report_nb_current: int = 0
-        self.threat_actor_batch_processor = self._create_threat_actor_batch_processor()
-        self.threat_actor_nb_current: int = 0
-        self.malware_family_batch_processor = (
-            self._create_malware_family_batch_processor()
-        )
-        self.malware_family_nb_current: int = 0
-
-    def _create_report_batch_processor(self) -> GenericBatchProcessor:
-        """Create and configure the batch processor.
-
-        Returns:
-            Configured GenericBatchProcessor instance
-
-        """
-        return GenericBatchProcessor(
-            work_manager=self.work_manager,
-            config=REPORT_BATCH_PROCESSOR_CONFIG,
-            logger=self.logger,
-        )
-
-    def _create_threat_actor_batch_processor(self) -> GenericBatchProcessor:
-        """Create and configure the threat actor batch processor.
-
-        Returns:
-            Configured GenericBatchProcessor instance
-
-        """
-        return GenericBatchProcessor(
-            work_manager=self.work_manager,
-            config=THREAT_ACTOR_BATCH_PROCESSOR_CONFIG,
-            logger=self.logger,
-        )
-
-    def _create_malware_family_batch_processor(self) -> GenericBatchProcessor:
-        """Create and configure the malware family batch processor.
-
-        Returns:
-            Configured GenericBatchProcessor instance
-
-        """
-        return GenericBatchProcessor(
-            work_manager=self.work_manager,
-            config=MALWARE_FAMILY_BATCH_PROCESSOR_CONFIG,
-            logger=self.logger,
-        )
+        if self.config.import_reports:
+            self.logger.info(
+                f"{LOG_PREFIX} Report import start date: {self.config.report_import_start_date}"
+            )
+            self.report_orchestrator = OrchestratorReport(
+                work_manager, logger, config, tlp_level
+            )
+        if self.config.import_threat_actors:
+            self.logger.info(
+                f"{LOG_PREFIX} Threat actor import start date: {self.config.threat_actor_import_start_date}"
+            )
+            self.threat_actor_orchestrator = OrchestratorThreatActor(
+                work_manager, logger, config, tlp_level
+            )
+        if self.config.import_malware_families:
+            self.logger.info(
+                f"{LOG_PREFIX} Malware family import start date: {self.config.malware_family_import_start_date}"
+            )
+            self.malware_orchestrator = OrchestratorMalware(
+                work_manager, logger, config, tlp_level
+            )
+        self.logger.info(f"{LOG_PREFIX} Orchestrator initialized")
 
     async def run_report(self, initial_state: Optional[Dict[str, Any]]) -> None:
-        """Run the orchestrator.
+        """Run the report orchestrator.
 
         Args:
             initial_state: Initial state for the orchestrator
 
         """
-        subentity_types = [
-            "malware_families",
-            "threat_actors",
-            "attack_techniques",
-            "vulnerabilities",
-            "domains",
-            "files",
-            "urls",
-            "ip_addresses",
-        ]
-        try:
-            async for gti_reports in self.client_api.fetch_reports(initial_state):
-                total_reports = len(gti_reports)
-                for report_idx, report in enumerate(gti_reports):
-                    report_entities = self.converter.convert_report_to_stix(report)
-                    subentities_ids = await self.client_api.fetch_subentities_ids(
-                        entity_name="entity_id",
-                        entity_id=report.id,
-                        subentity_types=subentity_types,
-                    )
-                    rel_summary = ", ".join(
-                        [f"{k}: {len(v)}" for k, v in subentities_ids.items()]
-                    )
-                    if len(rel_summary) > 0:
-                        self.logger.info(
-                            f"{LOG_PREFIX} ({report_idx + 1}/{total_reports}) Found relationships {{{rel_summary}}}"
-                        )
-                    subentities_detailed = (
-                        await self.client_api.fetch_subentity_details(subentities_ids)
-                    )
-                    subentity_stix = (
-                        self.converter.convert_subentities_to_stix_with_linking(
-                            subentities=subentities_detailed,
-                            main_entity="report",
-                            main_entities=report_entities,
-                        )
-                    )
-
-                    all_entities = report_entities + (subentity_stix or [])
-                    entity_types: Dict[str, int] = {}
-                    for entity in all_entities:
-                        entity_type = getattr(entity, "type", None)
-                        if entity_type:
-                            entity_types[entity_type] = (
-                                entity_types.get(entity_type, 0) + 1
-                            )
-                    entities_summary = ", ".join(
-                        [f"{k}: {v}" for k, v in entity_types.items()]
-                    )
-                    self.logger.info(
-                        f"{LOG_PREFIX} ({report_idx + 1}/{total_reports}) Converted to {len(all_entities)} STIX entities {{{entities_summary}}}"
-                    )
-                    if (
-                        self.report_batch_processor.get_current_batch_size()
-                        + len(all_entities)
-                    ) >= self.report_batch_processor.config.batch_size:
-                        self.logger.info(
-                            f"{LOG_PREFIX} Need to Flush before adding next items to preserve consistency of the bundle"
-                        )
-                        self.report_batch_processor.flush()
-                    self._update_report_index_inplace()
-                    self.report_batch_processor.add_item(self.converter.organization)
-                    self.report_batch_processor.add_item(self.converter.tlp_marking)
-                    self.report_batch_processor.add_items(all_entities)
-        finally:
-            self._flush_batch_processor()
+        self.logger.info(f"{LOG_PREFIX} Starting report orchestration")
+        await self.report_orchestrator.run(initial_state)
 
     async def run_threat_actor(self, initial_state: Optional[Dict[str, Any]]) -> None:
         """Run the threat actor orchestrator.
@@ -185,80 +89,8 @@ class Orchestrator:
             initial_state: Initial state for the orchestrator
 
         """
-        subentity_types = [
-            "malware_families",
-            "reports",
-            "attack_techniques",
-            "vulnerabilities",
-            #            "domains",
-            #            "files",
-            #            "urls",
-            #            "ip_addresses",
-        ]
-        try:
-            async for gti_threat_actors in self.client_api.fetch_threat_actors(
-                initial_state
-            ):
-                total_threat_actors = len(gti_threat_actors)
-                for threat_actor_idx, threat_actor in enumerate(gti_threat_actors):
-                    threat_actor_entities = self.converter.convert_threat_actor_to_stix(
-                        threat_actor
-                    )
-                    subentities_ids = await self.client_api.fetch_subentities_ids(
-                        entity_name="entity_id",
-                        entity_id=threat_actor.id,
-                        subentity_types=subentity_types,
-                    )
-                    rel_summary = ", ".join(
-                        [f"{k}: {len(v)}" for k, v in subentities_ids.items()]
-                    )
-                    if len(rel_summary) > 0:
-                        self.logger.info(
-                            f"{LOG_PREFIX} ({threat_actor_idx + 1}/{total_threat_actors}) Found relationships {{{rel_summary}}}"
-                        )
-                    subentities_detailed = (
-                        await self.client_api.fetch_subentity_details(subentities_ids)
-                    )
-                    subentity_stix = (
-                        self.converter.convert_subentities_to_stix_with_linking(
-                            subentities=subentities_detailed,
-                            main_entity="threat_actor",
-                            main_entities=threat_actor_entities,
-                        )
-                    )
-
-                    all_entities = threat_actor_entities + (subentity_stix or [])
-                    entity_types: Dict[str, int] = {}
-                    for entity in all_entities:
-                        entity_type = getattr(entity, "type", None)
-                        if entity_type:
-                            entity_types[entity_type] = (
-                                entity_types.get(entity_type, 0) + 1
-                            )
-                    entities_summary = ", ".join(
-                        [f"{k}: {v}" for k, v in entity_types.items()]
-                    )
-                    self.logger.info(
-                        f"{LOG_PREFIX} ({threat_actor_idx + 1}/{total_threat_actors}) Converted to {len(all_entities)} STIX entities {{{entities_summary}}}"
-                    )
-                    if (
-                        self.threat_actor_batch_processor.get_current_batch_size()
-                        + len(all_entities)
-                    ) >= self.threat_actor_batch_processor.config.batch_size:
-                        self.logger.info(
-                            f"{LOG_PREFIX} Need to Flush before adding next items to preserve consistency of the bundle"
-                        )
-                        self.threat_actor_batch_processor.flush()
-                    self._update_threat_actor_index_inplace()
-                    self.threat_actor_batch_processor.add_item(
-                        self.converter.organization
-                    )
-                    self.threat_actor_batch_processor.add_item(
-                        self.converter.tlp_marking
-                    )
-                    self.threat_actor_batch_processor.add_items(all_entities)
-        finally:
-            self._flush_threat_actor_batch_processor()
+        self.logger.info(f"{LOG_PREFIX} Starting threat actor orchestration")
+        await self.threat_actor_orchestrator.run(initial_state)
 
     async def run_malware_family(self, initial_state: Optional[Dict[str, Any]]) -> None:
         """Run the malware family orchestrator.
@@ -267,175 +99,5 @@ class Orchestrator:
             initial_state: Initial state for the orchestrator
 
         """
-        subentity_types = [
-            "threat_actors",
-            "reports",
-            "attack_techniques",
-            "vulnerabilities",
-            #            "domains",
-            #            "files",
-            #            "urls",
-            #            "ip_addresses",
-        ]
-        try:
-            async for gti_malware_families in self.client_api.fetch_malware_families(
-                initial_state
-            ):
-                total_malware_families = len(gti_malware_families)
-                for malware_family_idx, malware_family in enumerate(
-                    gti_malware_families
-                ):
-                    malware_family_entities = (
-                        self.converter.convert_malware_family_to_stix(malware_family)
-                    )
-                    subentities_ids = await self.client_api.fetch_subentities_ids(
-                        entity_name="entity_id",
-                        entity_id=malware_family.id,
-                        subentity_types=subentity_types,
-                    )
-                    rel_summary = ", ".join(
-                        [f"{k}: {len(v)}" for k, v in subentities_ids.items()]
-                    )
-                    if len(rel_summary) > 0:
-                        self.logger.info(
-                            f"{LOG_PREFIX} ({malware_family_idx + 1}/{total_malware_families}) Found relationships {{{rel_summary}}}"
-                        )
-                    subentities_detailed = (
-                        await self.client_api.fetch_subentity_details(subentities_ids)
-                    )
-                    subentity_stix = (
-                        self.converter.convert_subentities_to_stix_with_linking(
-                            subentities=subentities_detailed,
-                            main_entity="malware_family",
-                            main_entities=malware_family_entities,
-                        )
-                    )
-
-                    all_entities = malware_family_entities + (subentity_stix or [])
-                    entity_types: Dict[str, int] = {}
-                    for entity in all_entities:
-                        entity_type = getattr(entity, "type", None)
-                        if entity_type:
-                            entity_types[entity_type] = (
-                                entity_types.get(entity_type, 0) + 1
-                            )
-                    entities_summary = ", ".join(
-                        [f"{k}: {v}" for k, v in entity_types.items()]
-                    )
-                    self.logger.info(
-                        f"{LOG_PREFIX} ({malware_family_idx + 1}/{total_malware_families}) Converted to {len(all_entities)} STIX entities {{{entities_summary}}}"
-                    )
-                    if (
-                        self.malware_family_batch_processor.get_current_batch_size()
-                        + len(all_entities)
-                    ) >= self.malware_family_batch_processor.config.batch_size:
-                        self.logger.info(
-                            f"{LOG_PREFIX} Need to Flush before adding next items to preserve consistency of the bundle"
-                        )
-                        self.malware_family_batch_processor.flush()
-                    self._update_malware_family_index_inplace()
-                    self.malware_family_batch_processor.add_item(
-                        self.converter.organization
-                    )
-                    self.malware_family_batch_processor.add_item(
-                        self.converter.tlp_marking
-                    )
-                    self.malware_family_batch_processor.add_items(all_entities)
-        finally:
-            self._flush_malware_family_batch_processor()
-
-    def _update_report_index_inplace(self) -> None:
-        """Update the work message to reflect current report progress."""
-
-        def replacer(match: Any) -> str:
-            actual_total = self.client_api.real_total_reports or 0
-
-            if actual_total == 0:
-                return "(~ 0/0 reports)"
-
-            self.report_nb_current += 1
-            return f"(~ {self.report_nb_current}/{actual_total} reports)"
-
-        pattern = r"\(~ (\d+)/(\d+) reports\)"
-        template = self.report_batch_processor.config.work_name_template
-        self.report_batch_processor.config.work_name_template = re.sub(
-            pattern, replacer, template
-        )
-
-    def _update_threat_actor_index_inplace(self) -> None:
-        """Update the work message to reflect current threat actor progress."""
-
-        def replacer(match: Any) -> str:
-            actual_total = self.client_api.real_total_threat_actors or 0
-
-            if actual_total == 0:
-                return "(~ 0/0 threat actors)"
-
-            self.threat_actor_nb_current += 1
-            return f"(~ {self.threat_actor_nb_current}/{actual_total} threat actors)"
-
-        pattern = r"\(~ (\d+)/(\d+) threat actors\)"
-        template = self.threat_actor_batch_processor.config.work_name_template
-        self.threat_actor_batch_processor.config.work_name_template = re.sub(
-            pattern, replacer, template
-        )
-
-    def _update_malware_family_index_inplace(self) -> None:
-        """Update the work message to reflect current malware family progress."""
-
-        def replacer(match: Any) -> str:
-            actual_total = self.client_api.real_total_malware_families or 0
-
-            if actual_total == 0:
-                return "(~ 0/0 malware families)"
-
-            self.malware_family_nb_current += 1
-            return (
-                f"(~ {self.malware_family_nb_current}/{actual_total} malware families)"
-            )
-
-        pattern = r"\(~ (\d+)/(\d+) malware families\)"
-        template = self.malware_family_batch_processor.config.work_name_template
-        self.malware_family_batch_processor.config.work_name_template = re.sub(
-            pattern, replacer, template
-        )
-
-    def _flush_batch_processor(self) -> None:
-        """Flush any remaining items in the batch processor."""
-        try:
-            work_id = self.report_batch_processor.flush()
-            if work_id:
-                self.logger.info(
-                    f"{LOG_PREFIX} Batch processor: Flushed remaining items"
-                )
-            self.report_batch_processor.update_final_state()
-        except Exception as e:
-            self.logger.error(f"{LOG_PREFIX} Failed to flush batch processor: {str(e)}")
-
-    def _flush_threat_actor_batch_processor(self) -> None:
-        """Flush any remaining items in the threat actor batch processor."""
-        try:
-            work_id = self.threat_actor_batch_processor.flush()
-            if work_id:
-                self.logger.info(
-                    f"{LOG_PREFIX} Threat actor batch processor: Flushed remaining items"
-                )
-            self.threat_actor_batch_processor.update_final_state()
-        except Exception as e:
-            self.logger.error(
-                f"{LOG_PREFIX} Failed to flush threat actor batch processor: {str(e)}"
-            )
-
-    def _flush_malware_family_batch_processor(self) -> None:
-        """Flush any remaining items in the malware family batch processor."""
-        try:
-            work_id = self.malware_family_batch_processor.flush()
-            if work_id:
-                self.logger.info(
-                    f"{LOG_PREFIX} Malware family batch processor: Flushed remaining items"
-                )
-            self.malware_family_batch_processor.update_final_state()
-        except Exception as e:
-            self.logger.error(
-                f"{LOG_PREFIX} Failed to flush malware family batch processor: {str(e)}"
-            )
+        self.logger.info(f"{LOG_PREFIX} Starting malware family orchestration")
+        await self.malware_orchestrator.run(initial_state)
