+++ external-import/google-ti-feeds/tests/custom/test_orchestrator.py
@@ -179,9 +179,9 @@ def gti_config() -> DummyConfig:
 def expected_report_log_messages() -> list[str]:
     """Fixture for expected log messages in report orchestration."""
     return [
-        "[Fetchers] Fetched 1 reports from API (total of 1 items)",
-        "[Orchestrator] (1/1) Found relationships {malware_families: 1, threat_actors: 1, attack_techniques: 1, vulnerabilities: 1, domains: 1, files: 1, urls: 1, ip_addresses: 1}",
-        "[Fetchers] Fetching details for 8 subentities...",
+        "[FetcherReport] Fetched 1 reports from API (total of 1 items)",
+        "[OrchestratorReport] (1/1) Found relationships {malware_families: 1, threat_actors: 1, attack_techniques: 1, vulnerabilities: 1, domains: 1, files: 1, urls: 1, ip_addresses: 1}",
+        "[FetcherShared] Fetching details for 8 subentities...",
         "[GenericFetcher] Fetched 1 malware families",
         "[GenericFetcher] Fetched 1 threat actors",
         "[GenericFetcher] Fetched 1 attack techniques",
@@ -190,7 +190,7 @@ def expected_report_log_messages() -> list[str]:
         "[GenericFetcher] Fetched 1 files",
         "[GenericFetcher] Fetched 1 URLs",
         "[GenericFetcher] Fetched 1 IP addresses",
-        "[Fetchers] Fetched details {malware_families: 1, threat_actors: 1, attack_techniques: 1, vulnerabilities: 1, domains: 1, files: 1, urls: 1, ip_addresses: 1}",
+        "[FetcherShared] Fetched details {malware_families: 1, threat_actors: 1, attack_techniques: 1, vulnerabilities: 1, domains: 1, files: 1, urls: 1, ip_addresses: 1}",
         "[GenericConverter] Converted 33 malware families to STIX format",
         "[GenericConverter] Converted 53 threat actors to STIX format",
         "[GenericConverter] Converted 1 attack techniques to STIX format",
@@ -199,7 +199,7 @@ def expected_report_log_messages() -> list[str]:
         "[GenericConverter] Converted 3 files to STIX format",
         "[GenericConverter] Converted 3 URLs to STIX format",
         "[GenericConverter] Converted 3 IP addresses to STIX format",
-        "[Orchestrator] (1/1) Converted to 115 STIX entities {identity: 39, report: 1, malware: 1, relationship: 46, location: 17, intrusion-set: 1, attack-pattern: 1, vulnerability: 1, domain-name: 1, indicator: 4, file: 1, url: 1, ipv4-addr: 1}",
+        "[OrchestratorReport] (1/1) Converted to 115 STIX entities {identity: 39, report: 1, malware: 1, relationship: 46, location: 17, intrusion-set: 1, attack-pattern: 1, vulnerability: 1, domain-name: 1, indicator: 4, file: 1, url: 1, ipv4-addr: 1}",
         "[GenericBatchProcessor] Flushing remaining 117 STIX objects",
         "[GenericBatchProcessor] Processing batch #1 with 117 STIX objects (Total processed: 117)",
         "[GenericBatchProcessor] Sent batch #1 to OpenCTI",
@@ -213,19 +213,19 @@ def expected_report_log_messages() -> list[str]:
 def expected_threat_actor_log_messages() -> list[str]:
     """Fixture for expected log messages in threat actor orchestration."""
     return [
-        "[Fetchers] Fetched 1 threat_actors from API (total of 1 items)",
-        "[Orchestrator] (1/1) Found relationships {malware_families: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
-        "[Fetchers] Fetching details for 4 subentities...",
+        "[FetcherThreatActor] Fetched 1 threat_actors from API (total of 1 items)",
+        "[OrchestratorThreatActor] (1/1) Found relationships {malware_families: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
+        "[FetcherShared] Fetching details for 4 subentities...",
         "[GenericFetcher] Fetched 1 malware families",
         "[GenericFetcher] Fetched 1 reports",
         "[GenericFetcher] Fetched 1 attack techniques",
         "[GenericFetcher] Fetched 1 vulnerabilities",
-        "[Fetchers] Fetched details {malware_families: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
+        "[FetcherShared] Fetched details {malware_families: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
         "[GenericConverter] Converted 34 malware families to STIX format",
         "[GenericConverter] Converted 16 reports to STIX format",
         "[GenericConverter] Converted 2 attack techniques to STIX format",
         "[GenericConverter] Converted 2 vulnerabilities to STIX format",
-        "[Orchestrator] (1/1) Converted to 107 STIX entities {location: 17, identity: 39, intrusion-set: 1, relationship: 46, malware: 1, report: 1, attack-pattern: 1, vulnerability: 1}",
+        "[OrchestratorThreatActor] (1/1) Converted to 107 STIX entities {location: 17, identity: 39, intrusion-set: 1, relationship: 46, malware: 1, report: 1, attack-pattern: 1, vulnerability: 1}",
         "[GenericBatchProcessor] Flushing remaining 109 STIX objects",
         "[GenericBatchProcessor] Processing batch #1 with 109 STIX objects (Total processed: 109)",
         "[GenericBatchProcessor] Sent batch #1 to OpenCTI",
@@ -239,19 +239,19 @@ def expected_threat_actor_log_messages() -> list[str]:
 def expected_malware_family_log_messages() -> list[str]:
     """Fixture for expected log messages in malware family orchestration."""
     return [
-        "[Fetchers] Fetched 1 malware_families from API (total of 1 items)",
-        "[Orchestrator] (1/1) Found relationships {threat_actors: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
-        "[Fetchers] Fetching details for 4 subentities...",
+        "[FetcherMalware] Fetched 1 malware_families from API (total of 1 items)",
+        "[OrchestratorMalware] (1/1) Found relationships {threat_actors: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
+        "[FetcherShared] Fetching details for 4 subentities...",
         "[GenericFetcher] Fetched 1 threat actors",
         "[GenericFetcher] Fetched 1 reports",
         "[GenericFetcher] Fetched 1 attack techniques",
         "[GenericFetcher] Fetched 1 vulnerabilities",
-        "[Fetchers] Fetched details {threat_actors: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
+        "[FetcherShared] Fetched details {threat_actors: 1, reports: 1, attack_techniques: 1, vulnerabilities: 1}",
         "[GenericConverter] Converted 54 threat actors to STIX format",
         "[GenericConverter] Converted 16 reports to STIX format",
         "[GenericConverter] Converted 2 attack techniques to STIX format",
         "[GenericConverter] Converted 2 vulnerabilities to STIX format",
-        "[Orchestrator] (1/1) Converted to 107 STIX entities {identity: 39, malware: 1, relationship: 46, location: 17, intrusion-set: 1, report: 1, attack-pattern: 1, vulnerability: 1}",
+        "[OrchestratorMalware] (1/1) Converted to 107 STIX entities {identity: 39, malware: 1, relationship: 46, location: 17, intrusion-set: 1, report: 1, attack-pattern: 1, vulnerability: 1}",
         "[GenericBatchProcessor] Flushing remaining 109 STIX objects",
         "[GenericBatchProcessor] Processing batch #1 with 109 STIX objects (Total processed: 109)",
         "[GenericBatchProcessor] Sent batch #1 to OpenCTI",
